refactor: Rank 결과를 만드는 책임을 WinningLotto 클래스로 이동
