+++ src/main/kotlin/lotto/domain/WinningLotto.kt
@@ -4,6 +4,13 @@ class WinningLotto(
     val winningNumbers: Lotto,
     val bonusNumber: LottoNumber,
 ) {
+    fun matchLotto(targetLotto: Lotto): Rank {
+        return Rank.findByMatchCount(
+            countMatchingNumbers(targetLotto),
+            matchBonusNumber(targetLotto),
+        )
+    }
+
     fun countMatchingNumbers(targetLotto: Lotto): Int {
         return targetLotto.numbers.count { it in winningNumbers.numbers }
     }
